improved document loading times + updated benchmarks
